+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.kt
@@ -44,6 +44,10 @@ import at.hannibal2.skyhanni.features.minion.MinionCollectLogic
 import at.hannibal2.skyhanni.features.minion.MinionFeatures
 import at.hannibal2.skyhanni.features.misc.*
 import at.hannibal2.skyhanni.features.misc.discordrpc.DiscordRPCManager
+import at.hannibal2.skyhanni.features.misc.items.EstimatedItemValue
+import at.hannibal2.skyhanni.features.misc.items.EstimatedWardrobePrice
+import at.hannibal2.skyhanni.features.misc.teleportpad.TeleportPadCompactName
+import at.hannibal2.skyhanni.features.misc.teleportpad.TeleportPadInventoryNumber
 import at.hannibal2.skyhanni.features.misc.tiarelay.TiaRelayHelper
 import at.hannibal2.skyhanni.features.misc.tiarelay.TiaRelayWaypoints
 import at.hannibal2.skyhanni.features.misc.trevor.TrevorFeatures
@@ -94,7 +98,7 @@ import org.apache.logging.log4j.Logger
     clientSideOnly = true,
     useMetadata = true,
     guiFactory = "at.hannibal2.skyhanni.config.ConfigGuiForgeInterop",
-    version = "0.18.Beta.5",
+    version = "0.18.Beta.7",
 )
 class SkyHanniMod {
     @Mod.EventHandler
@@ -245,19 +249,20 @@ class SkyHanniMod {
         loadModule(FarmingArmorDrops())
         loadModule(JoinCrystalHollows())
         loadModule(GardenVisitorColorNames())
-        loadModule(GardenTeleportPadCompactName())
+        loadModule(TeleportPadCompactName())
         loadModule(AnitaMedalProfit())
         loadModule(ComposterDisplay())
         loadModule(GardenComposterInventoryFeatures())
         loadModule(MinionCollectLogic())
         loadModule(PasteIntoSigns())
-        loadModule(EstimatedItemValue())
+        loadModule(EstimatedItemValue)
+        loadModule(EstimatedWardrobePrice())
         loadModule(ComposterInventoryNumbers())
         loadModule(FarmingFortuneDisplay())
         loadModule(ToolTooltipTweaks)
         loadModule(CropSpeedMeter())
         loadModule(AshfangMinisNametagHider())
-        loadModule(GardenTeleportPadInventoryNumber())
+        loadModule(TeleportPadInventoryNumber())
         loadModule(ComposterOverlay())
         loadModule(DiscordRPCManager)
         loadModule(GardenCropMilestoneFix())
@@ -268,7 +273,6 @@ class SkyHanniMod {
         loadModule(GardenYawAndPitch())
         loadModule(MovementSpeedDisplay())
         loadModule(ChumBucketHider())
-        loadModule(GardenRecentTeleportPadsDisplay())
         loadModule(InquisitorWaypointShare)
         loadModule(TrevorFeatures())
         loadModule(TrevorSolver)
@@ -276,6 +280,8 @@ class SkyHanniMod {
         loadModule(GardenVisitorDropStatistics)
         loadModule(CaptureFarmingGear())
         loadModule(SackDisplay())
+        loadModule(GardenStartLocation)
+        loadModule(PetCandyUsedDisplay())
 
         init()