+++ src/main/java/at/hannibal2/skyhanni/config/Storage.java
@@ -154,6 +154,10 @@ public class Storage {
             }
 
             @Expose
+            public Map<CropType, LorenzVec> cropStartLocations = new HashMap<>();
+        }
+
+            @Expose
             public Fortune fortune = new Fortune();
 
             public static class Fortune {