+++ src/main/java/at/hannibal2/skyhanni/features/fishing/OdgerWaypoint.kt
@@ -3,13 +3,13 @@ package at.hannibal2.skyhanni.features.fishing
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.data.IslandType
 import at.hannibal2.skyhanni.test.GriffinUtils.drawWaypointFilled
+import at.hannibal2.skyhanni.utils.InventoryUtils
 import at.hannibal2.skyhanni.utils.ItemUtils.getLore
 import at.hannibal2.skyhanni.utils.ItemUtils.name
 import at.hannibal2.skyhanni.utils.LorenzColor
 import at.hannibal2.skyhanni.utils.LorenzUtils
 import at.hannibal2.skyhanni.utils.LorenzVec
 import at.hannibal2.skyhanni.utils.RenderUtils.drawDynamicText
-import net.minecraft.client.Minecraft
 import net.minecraftforge.client.event.RenderWorldLastEvent
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import net.minecraftforge.fml.common.gameevent.TickEvent
@@ -24,24 +24,19 @@ class OdgerWaypoint {
     @SubscribeEvent
     fun onTick(event: TickEvent.ClientTickEvent) {
         if (event.phase != TickEvent.Phase.START) return
-
-        if (!LorenzUtils.inSkyBlock) return
-        if (!SkyHanniMod.feature.fishing.odgerLocation) return
+        if (!isEnabled()) return
 
         tick++
-
         if (tick % 10 == 0) {
             hasLavaRodInHand = isLavaFishingRod()
 
-            trophyFishInInv = Minecraft.getMinecraft().thePlayer.inventoryContainer.inventorySlots
-                .mapNotNull { it?.stack }
-                .map { it.getLore() }
+            trophyFishInInv = InventoryUtils.getItemsInOwnInventory().map { it.getLore() }
                 .any { it.isNotEmpty() && it.last().endsWith("TROPHY FISH") }
         }
     }
 
     private fun isLavaFishingRod(): Boolean {
-        val heldItem = Minecraft.getMinecraft().thePlayer.heldItem ?: return false
+        val heldItem = InventoryUtils.getItemInHand() ?: return false
         val isRod = heldItem.name?.contains("Rod") ?: return false
         if (!isRod) return false
 
@@ -50,13 +45,14 @@ class OdgerWaypoint {
 
     @SubscribeEvent
     fun onRenderWorld(event: RenderWorldLastEvent) {
-        if (!LorenzUtils.inSkyBlock) return
-        if (LorenzUtils.skyBlockIsland != IslandType.CRIMSON_ISLE) return
-        if (!SkyHanniMod.feature.fishing.odgerLocation) return
+        if (!isEnabled()) return
         if (hasLavaRodInHand) return
         if (!trophyFishInInv) return
 
         event.drawWaypointFilled(location, LorenzColor.WHITE.toColor())
         event.drawDynamicText(location, "Odger", 1.5)
     }
+
+    fun isEnabled() = LorenzUtils.inSkyBlock && SkyHanniMod.feature.fishing.odgerLocation &&
+            LorenzUtils.skyBlockIsland == IslandType.CRIMSON_ISLE
 }