+++ src/main/java/at/hannibal2/skyhanni/features/misc/teleportpad/TeleportPadCompactName.kt
@@ -1,6 +1,8 @@
-package at.hannibal2.skyhanni.features.garden
+package at.hannibal2.skyhanni.features.misc.teleportpad
 
 import at.hannibal2.skyhanni.SkyHanniMod
+import at.hannibal2.skyhanni.data.IslandType
+import at.hannibal2.skyhanni.utils.LorenzUtils
 import at.hannibal2.skyhanni.utils.StringUtils.matchMatcher
 import net.minecraft.entity.EntityLivingBase
 import net.minecraft.entity.item.EntityArmorStand
@@ -8,13 +10,14 @@ import net.minecraftforge.client.event.RenderLivingEvent
 import net.minecraftforge.fml.common.eventhandler.EventPriority
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 
-class GardenTeleportPadCompactName {
+class TeleportPadCompactName {
     private val patternName = "§.✦ §aWarp To (?<name>.*)".toPattern()
     private val patternNoName = "§.✦ §cNo Destination".toPattern()
 
     @SubscribeEvent(priority = EventPriority.HIGH)
     fun onRenderLivingB(event: RenderLivingEvent.Specials.Pre<EntityLivingBase>) {
-        if (!SkyHanniMod.feature.garden.teleportPadsCompactName) return
+        if (!SkyHanniMod.feature.misc.teleportPad.compactName) return
+        if (LorenzUtils.skyBlockIsland != IslandType.PRIVATE_ISLAND) return
         val entity = event.entity
         if (entity !is EntityArmorStand) return