+++ src/main/java/at/hannibal2/skyhanni/utils/InventoryUtils.kt
@@ -3,23 +3,11 @@ package at.hannibal2.skyhanni.utils
 import net.minecraft.client.Minecraft
 import net.minecraft.client.gui.inventory.GuiChest
 import net.minecraft.inventory.ContainerChest
-import net.minecraft.inventory.Slot
 import net.minecraft.item.ItemStack
 
 object InventoryUtils {
 
-    //TODO use this method more widely
-    fun currentlyOpenInventory(): String {
-        val screen = Minecraft.getMinecraft().currentScreen
-        if (screen !is GuiChest) return ""
-        val inventorySlots = screen.inventorySlots
-        val chest = inventorySlots as ContainerChest
-
-        return chest.getInventoryName()
-    }
-
-    fun getItemsInOpenChest(): List<Slot> {
-        val list = mutableListOf<Slot>()
+    fun getItemsInOpenChest() = buildList {
         val guiChest = Minecraft.getMinecraft().currentScreen as GuiChest
         val inventorySlots = guiChest.inventorySlots.inventorySlots
         val skipAt = inventorySlots.size - 9 * 4
@@ -27,45 +15,31 @@ object InventoryUtils {
         for (slot in inventorySlots) {
             val stack = slot.stack
             if (stack != null) {
-                list.add(slot)
+                add(slot)
             }
             i++
             if (i == skipAt) break
         }
-        return list
     }
 
-    fun openInventoryName(): String {
-        val guiChest = Minecraft.getMinecraft().currentScreen
-        val chestName = if (guiChest is GuiChest) {
-            val chest = guiChest.inventorySlots as ContainerChest
-            chest.getInventoryName()
-        } else ""
-        return chestName
-    }
+    fun openInventoryName() = Minecraft.getMinecraft().currentScreen.let {
+         if (it is GuiChest) {
+             val chest = it.inventorySlots as ContainerChest
+             chest.getInventoryName()
+         } else ""
+     }
 
     fun ContainerChest.getInventoryName() = this.lowerChestInventory.displayName.unformattedText.trim()
 
-    fun getItemsInOwnInventory(): MutableList<ItemStack> {
-        val list = mutableListOf<ItemStack>()
-        for (itemStack in Minecraft.getMinecraft().thePlayer.inventory.mainInventory) {
-            itemStack?.let {
-                list.add(it)
-            }
-        }
+    fun getItemsInOwnInventory() = Minecraft.getMinecraft().thePlayer.inventory.mainInventory.filterNotNull()
 
-        return list
-    }
-
-    fun countItemsInLowerInventory(ignoreStackSize: Boolean = false, predicate: (ItemStack) -> Boolean) =
-        getItemsInOwnInventory().filter { predicate(it) }.sumOf { if (!ignoreStackSize) it.stackSize else 1 }
-
-    fun countItemsInOpenChest(ignoreStackSize: Boolean = false, predicate: (ItemStack) -> Boolean) =
-        getItemsInOpenChest().mapNotNull { it.stack }.filter { predicate(it) }
-            .sumOf { if (!ignoreStackSize) it.stackSize else 1 }
+    fun countItemsInLowerInventory(predicate: (ItemStack) -> Boolean) =
+        getItemsInOwnInventory().filter { predicate(it) }.sumOf { it.stackSize }
 
     fun getArmor(): Array<ItemStack?> = Minecraft.getMinecraft().thePlayer.inventory.armorInventory
 
     fun inStorage() =
         openInventoryName().let { it.contains("Storage") || it.contains("Ender Chest") || it.contains("Backpack") }
+
+    fun getItemInHand(): ItemStack? = Minecraft.getMinecraft().thePlayer.heldItem
 }